stock_tracker/requirements.txt
djangorestframework
django
djangorestframework-simplejwt
django-cors-headers
bcrypt
django-allauth
django-axes
django-rest-passwordreset
django-environ
